% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print_imath_version}
\alias{print_imath_version}
\title{Print the Imath library version info}
\usage{
print_imath_version()
}
\value{
None.
}
\description{
Print the Imath library version info
}
\examples{
# Print the Imath version provided in the static library
print_imath_version()
}
